+++ desktop/src/main/java/com/onecat/burst/world/WorldGizmo.java
@@ -5,6 +5,8 @@ import net.mgsx.gltf.scene3d.scene.SceneModel;
 
 public class WorldGizmo extends Scene {
 
+	public boolean visible = true;
+
 	public WorldGizmo(SceneModel sceneModel) {
 		super(sceneModel);
 	}