+++ docs/changelog.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Controllers management, atlas loading
+**Sep 2, 2026**
+- Consolidated all assets in assets folder root (there are not going to be as many enough assets to separate them into subfolders anyway)
+- Removed test PFX from repo
+- Set min/max size limits for Desktop window
+- Editor settings are functional now and are properly saved in Preferences
+- Effects can now be saved to a PFX file
+- Controllers can be cloned, removed, added and made invisible now
+- Fixed the issue Flame had about losing information on .atlas file used in the particle system upon loading and saving it
+- Atlas files can now be loaded and used by effects
+- A default particle effect is now created upon session start or creating a new file
+  -Made Settings entries and defaults statically typed and centralized in Settings class
+- Updated README with miserable explanation why even bother use Burst instead of Flame
+
 ## Camera, PFX loading, Settings
 **Aug 30, 2026**
 - Made gizmo a bit thinner